backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
pydantic-settings
